PR 3: Extract LLMService from Agent
