+++ browser_use/agent/service.py
@@ -22,15 +22,15 @@ from browser_use.agent.cloud_events import (
 	CreateAgentTaskEvent,
 	UpdateAgentTaskEvent,
 )
+from browser_use.agent.llm_service import LLMService
 from browser_use.llm.base import BaseChatModel
-from browser_use.llm.exceptions import ModelOutputTruncatedError, ModelProviderError, ModelRateLimitError
-from browser_use.llm.messages import BaseMessage, ContentPartImageParam, ContentPartTextParam, UserMessage
+from browser_use.llm.messages import BaseMessage, ContentPartImageParam, ContentPartTextParam
 from browser_use.tokens.service import TokenCost
 
 load_dotenv()
 
 from bubus import EventBus
-from pydantic import BaseModel, ValidationError
+from pydantic import BaseModel
 from uuid_extensions import uuid7str
 
 from browser_use import Browser, BrowserProfile, BrowserSession
@@ -76,7 +76,6 @@ from browser_use.telemetry.views import AgentTelemetryEvent
 from browser_use.tools.registry.views import ActionModel
 from browser_use.tools.service import Tools
 from browser_use.utils import (
-	URL_PATTERN,
 	_log_pretty_path,
 	check_latest_browser_use_version,
 	get_browser_use_version,
@@ -89,48 +88,8 @@ from browser_use.utils import (
 logger = logging.getLogger(__name__)
 
 
-def log_response(response: AgentOutput, registry=None, logger=None) -> None:
-	"""Utility function to log the model's response."""
-
-	# Use module logger if no logger provided
-	if logger is None:
-		logger = logging.getLogger(__name__)
-
-	# Only log thinking if it's present
-	if response.current_state.thinking:
-		logger.debug(f'💡 Thinking:\n{response.current_state.thinking}')
-
-	# Only log evaluation if it's not empty
-	eval_goal = response.current_state.evaluation_previous_goal
-	if eval_goal:
-		if 'success' in eval_goal.lower():
-			emoji = '👍'
-			# Green color for success
-			logger.info(f'  \033[32m{emoji} Eval: {eval_goal}\033[0m')
-		elif 'failure' in eval_goal.lower():
-			emoji = '⚠️'
-			# Red color for failure
-			logger.info(f'  \033[31m{emoji} Eval: {eval_goal}\033[0m')
-		else:
-			emoji = '❔'
-			# No color for unknown/neutral
-			logger.info(f'  {emoji} Eval: {eval_goal}')
-
-	# Always log memory if present
-	if response.current_state.memory:
-		logger.info(f'  🧠 Memory: {response.current_state.memory}')
-
-	# Only log next goal if it's not empty
-	next_goal = response.current_state.next_goal
-	if next_goal:
-		# Blue color for next goal
-		logger.info(f'  \033[34m🎯 Next goal: {next_goal}\033[0m')
-
-
-Context = TypeVar('Context')
-
-
 AgentHookFunc = Callable[['Agent'], Awaitable[None]]
+Context = TypeVar('Context')
 
 
 class Agent(Generic[Context, AgentStructuredOutput]):
@@ -380,10 +339,6 @@ class Agent(Generic[Context, AgentStructuredOutput]):
 		self.llm = llm
 		self.judge_llm = judge_llm
 
-		# Fallback LLM configuration
-		self._fallback_llm: BaseChatModel | None = fallback_llm
-		self._using_fallback_llm: bool = False
-		self._original_llm: BaseChatModel = llm  # Store original for reference
 		self.directly_open_url = directly_open_url
 		self.include_recent_events = include_recent_events
 		self._url_shortening_limit = _url_shortening_limit
@@ -478,8 +433,6 @@ class Agent(Generic[Context, AgentStructuredOutput]):
 		self.initial_url = initial_url
 
 		self.initial_actions = self._convert_initial_actions(initial_actions) if initial_actions else None
-		# Verify we can connect to the model
-		self._verify_and_setup_llm()
 
 		# TODO: move this logic to the LLMs
 		# Handle users trying to use use_vision=True with DeepSeek models
@@ -547,6 +500,10 @@ class Agent(Generic[Context, AgentStructuredOutput]):
 			action_phase=action_phase,
 			post_processor=post_processor,
 		)
+		# Verify we can connect to the model
+		self._llm_service = LLMService(self, self._pipeline)
+		self._llm_service.setup_fallback(fallback_llm)
+		self._llm_service.verify_and_setup_llm()
 		if self.sensitive_data:
 			# Check if sensitive_data has domain-specific credentials
 			has_domain_specific_credentials = any(isinstance(v, dict) for v in self.sensitive_data.values())
@@ -664,12 +621,12 @@ class Agent(Generic[Context, AgentStructuredOutput]):
 	@property
 	def is_using_fallback_llm(self) -> bool:
 		"""Check if the agent is currently using the fallback LLM."""
-		return self._using_fallback_llm
+		return self._llm_service.is_using_fallback_llm
 
 	@property
 	def current_llm_model(self) -> str:
 		"""Get the model name of the currently active LLM."""
-		return self.llm.model if hasattr(self.llm, 'model') else 'unknown'
+		return self._llm_service.current_llm_model
 
 	def _set_file_system(self, file_system_path: str | None = None) -> None:
 		# Check for conflicting parameters
@@ -786,7 +743,6 @@ class Agent(Generic[Context, AgentStructuredOutput]):
 			"[Cloned] Github Stars Tracker" -> "cloned_github_stars_tracker"
 			"Get Weather Data" -> "get_weather_data"
 		"""
-		import re
 
 		# Remove special characters and convert to lowercase
 		slug = re.sub(r'[^\w\s]', '', skill.title.lower())
@@ -1191,147 +1147,7 @@ class Agent(Generic[Context, AgentStructuredOutput]):
 				judge_log += f'   {judgement.reasoning}\n'
 				self.logger.info(judge_log)
 
-	async def _get_model_output_with_retry(self, input_messages: list[BaseMessage]) -> AgentOutput:
-		"""Get model output with retry logic for empty actions"""
-		model_output = await self.get_model_output(input_messages)
-		self.logger.debug(
-			f'✅ Step {self.state.n_steps}: Got LLM response with {len(model_output.action) if model_output.action else 0} actions'
-		)
-
-		if (
-			not model_output.action
-			or not isinstance(model_output.action, list)
-			or all(action.model_dump() == {} for action in model_output.action)
-		):
-			self.logger.warning('Model returned empty action. Retrying...')
-
-			clarification_message = UserMessage(
-				content='You forgot to return an action. Please respond with a valid JSON action according to the expected schema with your assessment and next actions.'
-			)
-
-			retry_messages = input_messages + [clarification_message]
-			model_output = await self.get_model_output(retry_messages)
-
-			if not model_output.action or all(action.model_dump() == {} for action in model_output.action):
-				self.logger.warning('Model still returned empty after retry. Inserting safe noop action.')
-				action_instance = self.ActionModel()
-				setattr(
-					action_instance,
-					'done',
-					{
-						'success': False,
-						'text': 'No next action returned by LLM!',
-					},
-				)
-				model_output.action = [action_instance]
-
-		return model_output
-
-	def _remove_think_tags(self, text: str) -> str:
-		THINK_TAGS = re.compile(r'<think>.*?</think>', re.DOTALL)
-		STRAY_CLOSE_TAG = re.compile(r'.*?</think>', re.DOTALL)
-		# Step 1: Remove well-formed <think>...</think>
-		text = re.sub(THINK_TAGS, '', text)
-		# Step 2: If there's an unmatched closing tag </think>,
-		#         remove everything up to and including that.
-		text = re.sub(STRAY_CLOSE_TAG, '', text)
-		return text.strip()
-
-	# region - URL replacement
-	def _replace_urls_in_text(self, text: str) -> tuple[str, dict[str, str]]:
-		"""Replace URLs in a text string"""
-
-		replaced_urls: dict[str, str] = {}
-
-		def replace_url(match: re.Match) -> str:
-			"""Url can only have 1 query and 1 fragment"""
-			import hashlib
-
-			original_url = match.group(0)
-
-			# Find where the query/fragment starts
-			query_start = original_url.find('?')
-			fragment_start = original_url.find('#')
-
-			# Find the earliest position of query or fragment
-			after_path_start = len(original_url)  # Default: no query/fragment
-			if query_start != -1:
-				after_path_start = min(after_path_start, query_start)
-			if fragment_start != -1:
-				after_path_start = min(after_path_start, fragment_start)
-
-			# Split URL into base (up to path) and after_path (query + fragment)
-			base_url = original_url[:after_path_start]
-			after_path = original_url[after_path_start:]
-
-			# If after_path is within the limit, don't shorten
-			if len(after_path) <= self._url_shortening_limit:
-				return original_url
-
-			# If after_path is too long, truncate and add hash
-			if after_path:
-				truncated_after_path = after_path[: self._url_shortening_limit]
-				# Create a short hash of the full after_path content
-				hash_obj = hashlib.md5(after_path.encode('utf-8'))
-				short_hash = hash_obj.hexdigest()[:7]
-				# Create shortened URL
-				shortened = f'{base_url}{truncated_after_path}...{short_hash}'
-				# Only use shortened URL if it's actually shorter than the original
-				if len(shortened) < len(original_url):
-					replaced_urls[shortened] = original_url
-					return shortened
-
-			return original_url
-
-		return URL_PATTERN.sub(replace_url, text), replaced_urls
-
-	def _process_messsages_and_replace_long_urls_shorter_ones(self, input_messages: list[BaseMessage]) -> dict[str, str]:
-		"""Replace long URLs with shorter ones
-		? @dev edits input_messages in place
-
-		returns:
-			tuple[filtered_input_messages, urls we replaced {shorter_url: original_url}]
-		"""
-		from browser_use.llm.messages import AssistantMessage, UserMessage
-
-		urls_replaced: dict[str, str] = {}
-
-		# Process each message, in place
-		for message in input_messages:
-			# no need to process SystemMessage, we have control over that anyway
-			if isinstance(message, (UserMessage, AssistantMessage)):
-				if isinstance(message.content, str):
-					# Simple string content
-					message.content, replaced_urls = self._replace_urls_in_text(message.content)
-					urls_replaced.update(replaced_urls)
-
-				elif isinstance(message.content, list):
-					# List of content parts
-					for part in message.content:
-						if isinstance(part, ContentPartTextParam):
-							part.text, replaced_urls = self._replace_urls_in_text(part.text)
-							urls_replaced.update(replaced_urls)
-
-		return urls_replaced
-
 	@staticmethod
-	def _recursive_process_all_strings_inside_pydantic_model(model: BaseModel, url_replacements: dict[str, str]) -> None:
-		"""Recursively process all strings inside a Pydantic model, replacing shortened URLs with originals in place."""
-		for field_name, field_value in model.__dict__.items():
-			if isinstance(field_value, str):
-				# Replace shortened URLs with original URLs in string
-				processed_string = Agent._replace_shortened_urls_in_string(field_value, url_replacements)
-				setattr(model, field_name, processed_string)
-			elif isinstance(field_value, BaseModel):
-				# Recursively process nested Pydantic models
-				Agent._recursive_process_all_strings_inside_pydantic_model(field_value, url_replacements)
-			elif isinstance(field_value, dict):
-				# Process dictionary values in place
-				Agent._recursive_process_dict(field_value, url_replacements)
-			elif isinstance(field_value, (list, tuple)):
-				processed_value = Agent._recursive_process_list_or_tuple(field_value, url_replacements)
-				setattr(model, field_name, processed_value)
-
 	@staticmethod
 	def _recursive_process_dict(dictionary: dict, url_replacements: dict[str, str]) -> None:
 		"""Helper method to process dictionaries."""
@@ -1339,7 +1155,7 @@ class Agent(Generic[Context, AgentStructuredOutput]):
 			if isinstance(v, str):
 				dictionary[k] = Agent._replace_shortened_urls_in_string(v, url_replacements)
 			elif isinstance(v, BaseModel):
-				Agent._recursive_process_all_strings_inside_pydantic_model(v, url_replacements)
+				LLMService._recursive_process_all_strings_inside_pydantic_model(v, url_replacements)
 			elif isinstance(v, dict):
 				Agent._recursive_process_dict(v, url_replacements)
 			elif isinstance(v, (list, tuple)):
@@ -1355,7 +1171,7 @@ class Agent(Generic[Context, AgentStructuredOutput]):
 				if isinstance(item, str):
 					processed_items.append(Agent._replace_shortened_urls_in_string(item, url_replacements))
 				elif isinstance(item, BaseModel):
-					Agent._recursive_process_all_strings_inside_pydantic_model(item, url_replacements)
+					LLMService._recursive_process_all_strings_inside_pydantic_model(item, url_replacements)
 					processed_items.append(item)
 				elif isinstance(item, dict):
 					Agent._recursive_process_dict(item, url_replacements)
@@ -1371,7 +1187,7 @@ class Agent(Generic[Context, AgentStructuredOutput]):
 				if isinstance(item, str):
 					container[i] = Agent._replace_shortened_urls_in_string(item, url_replacements)
 				elif isinstance(item, BaseModel):
-					Agent._recursive_process_all_strings_inside_pydantic_model(item, url_replacements)
+					LLMService._recursive_process_all_strings_inside_pydantic_model(item, url_replacements)
 				elif isinstance(item, dict):
 					Agent._recursive_process_dict(item, url_replacements)
 				elif isinstance(item, (list, tuple)):
@@ -1389,101 +1205,6 @@ class Agent(Generic[Context, AgentStructuredOutput]):
 	# endregion - URL replacement
 
 	@time_execution_async('--get_next_action')
-	@observe_debug(ignore_input=True, ignore_output=True, name='get_model_output')
-	async def get_model_output(self, input_messages: list[BaseMessage]) -> AgentOutput:
-		"""Get next action from LLM based on current state"""
-
-		urls_replaced = self._process_messsages_and_replace_long_urls_shorter_ones(input_messages)
-
-		# Build kwargs for ainvoke
-		# Note: ChatBrowserUse will automatically generate action descriptions from output_format schema
-		kwargs: dict = {'output_format': self.AgentOutput, 'session_id': self.session_id}
-
-		try:
-			response = await self.llm.ainvoke(input_messages, **kwargs)
-			parsed: AgentOutput = response.completion  # type: ignore[assignment]
-
-			# Replace any shortened URLs in the LLM response back to original URLs
-			if urls_replaced:
-				self._recursive_process_all_strings_inside_pydantic_model(parsed, urls_replaced)
-
-			# cut the number of actions to max_actions_per_step if needed
-			if len(parsed.action) > self.settings.max_actions_per_step:
-				parsed.action = parsed.action[: self.settings.max_actions_per_step]
-
-			if not (hasattr(self.state, 'paused') and (self.state.paused or self.state.stopped)):
-				log_response(parsed, self.tools.registry.registry, self.logger)
-				await self._pipeline.broadcast_model_state(parsed)
-
-			self._log_next_action_summary(parsed)
-			return parsed
-		except ValidationError:
-			# Just re-raise - Pydantic's validation errors are already descriptive
-			raise
-		except (ModelRateLimitError, ModelProviderError) as e:
-			# Check if we can switch to a fallback LLM
-			if not self._try_switch_to_fallback_llm(e):
-				# No fallback available, re-raise the original error
-				raise
-			# Retry with the fallback LLM
-			return await self.get_model_output(input_messages)
-
-	def _try_switch_to_fallback_llm(self, error: ModelRateLimitError | ModelProviderError) -> bool:
-		"""
-		Attempt to switch to a fallback LLM after a rate limit or provider error.
-
-		Returns True if successfully switched to a fallback, False if no fallback available.
-		Once switched, the agent will use the fallback LLM for the rest of the run.
-		"""
-		# Already using fallback - can't switch again
-		if self._using_fallback_llm:
-			self.logger.warning(
-				f'⚠️ Fallback LLM also failed ({type(error).__name__}: {error.message}), no more fallbacks available'
-			)
-			return False
-
-		# Check if error is retryable (rate limit, auth errors, or server errors)
-		# 401: API key invalid/expired - fallback to different provider
-		# 402: Insufficient credits/payment required - fallback to different provider
-		# 429: Rate limit exceeded
-		# 500, 502, 503, 504: Server errors
-		# ModelOutputTruncatedError: not retryable on the same model, but a fallback may have a higher cap
-		retryable_status_codes = {401, 402, 429, 500, 502, 503, 504}
-		is_retryable = isinstance(error, (ModelRateLimitError, ModelOutputTruncatedError)) or (
-			hasattr(error, 'status_code') and error.status_code in retryable_status_codes
-		)
-
-		if not is_retryable:
-			return False
-
-		# Check if we have a fallback LLM configured
-		if self._fallback_llm is None:
-			self.logger.warning(f'⚠️ LLM error ({type(error).__name__}: {error.message}) but no fallback_llm configured')
-			return False
-
-		self._log_fallback_switch(error, self._fallback_llm)
-
-		# Switch to the fallback LLM
-		self.llm = self._fallback_llm
-		self._using_fallback_llm = True
-
-		# Register the fallback LLM for token cost tracking
-		self.token_cost_service.register_llm(self._fallback_llm)
-
-		return True
-
-	def _log_fallback_switch(self, error: ModelRateLimitError | ModelProviderError, fallback: BaseChatModel) -> None:
-		"""Log when switching to a fallback LLM."""
-		original_model = self._original_llm.model if hasattr(self._original_llm, 'model') else 'unknown'
-		fallback_model = fallback.model if hasattr(fallback, 'model') else 'unknown'
-		error_type = type(error).__name__
-		status_code = getattr(error, 'status_code', 'N/A')
-
-		self.logger.warning(
-			f'⚠️ Primary LLM ({original_model}) failed with {error_type} (status={status_code}), '
-			f'switching to fallback LLM ({fallback_model})'
-		)
-
 	async def _log_agent_run(self) -> None:
 		"""Log the agent run"""
 		# Blue color for task
@@ -1505,38 +1226,6 @@ class Agent(Generic[Context, AgentStructuredOutput]):
 			self.logger.info(
 				f'Starting a browser-use agent with version {self.version}, with provider={self.llm.provider} and model={self.llm.model}'
 			)
-
-	def _log_next_action_summary(self, parsed: 'AgentOutput') -> None:
-		"""Log a comprehensive summary of the next action(s)"""
-		if not (self.logger.isEnabledFor(logging.DEBUG) and parsed.action):
-			return
-
-		# Collect action details
-		action_details = []
-		for i, action in enumerate(parsed.action):
-			action_data = action.model_dump(exclude_unset=True)
-			action_name = next(iter(action_data.keys())) if action_data else 'unknown'
-			action_params = action_data.get(action_name, {}) if action_data else {}
-
-			# Format key parameters concisely
-			param_summary = []
-			if isinstance(action_params, dict):
-				for key, value in action_params.items():
-					if key == 'index':
-						param_summary.append(f'#{value}')
-					elif key == 'text' and isinstance(value, str):
-						text_preview = value[:30] + '...' if len(value) > 30 else value
-						param_summary.append(f'text="{text_preview}"')
-					elif key == 'url':
-						param_summary.append(f'url="{value}"')
-					elif key == 'success':
-						param_summary.append(f'success={value}')
-					elif isinstance(value, (str, int, bool)):
-						val_str = str(value)[:30] + '...' if len(str(value)) > 30 else str(value)
-						param_summary.append(f'{key}={val_str}')
-
-			param_str = f'({", ".join(param_summary)})' if param_summary else ''
-			action_details.append(f'{action_name}{param_str}')
 
 	def _log_final_outcome_messages(self) -> None:
 		"""Log helpful messages to user based on agent run outcome"""
@@ -1665,8 +1354,6 @@ class Agent(Generic[Context, AgentStructuredOutput]):
 
 	def _extract_start_url(self, task: str) -> str | None:
 		"""Extract URL from task string using naive pattern matching."""
-
-		import re
 
 		# Remove email addresses from task before looking for URLs
 		task_without_emails = re.sub(r'\b[A-Za-z0-9._%+-]+@[A-Za-z0-9.-]+\.[A-Z|a-z]{2,}\b', '', task)
@@ -2180,17 +1867,6 @@ class Agent(Generic[Context, AgentStructuredOutput]):
 
 		return converted_actions
 
-	def _verify_and_setup_llm(self):
-		"""
-		Verify that the LLM API keys are setup and the LLM API is responding properly.
-		Also handles tool calling method detection if in auto mode.
-		"""
-
-		# Skip verification if already done
-		if getattr(self.llm, '_verified_api_keys', None) is True or CONFIG.SKIP_LLM_API_KEY_VERIFICATION:
-			setattr(self.llm, '_verified_api_keys', True)
-			return True
-
 	@property
 	def message_manager(self) -> MessageManager:
 		return self._message_manager
@@ -2378,6 +2054,20 @@ class Agent(Generic[Context, AgentStructuredOutput]):
 					elif isinstance(item, dict):
 						count += self._substitute_in_dict(item, replacements)
 		return count
+
+	# ── LLMService forwarding wrappers ──────────────────────────
+
+	async def get_model_output(self, input_messages: list[BaseMessage]) -> AgentOutput:
+		"""Get next action from LLM based on current state (delegates to LLMService)."""
+		return await self._llm_service.get_model_output(input_messages)
+
+	async def _get_model_output_with_retry(self, input_messages: list[BaseMessage]) -> AgentOutput:
+		"""Get model output with retry logic (delegates to LLMService)."""
+		return await self._llm_service.get_model_output_with_retry(input_messages)
+
+	def _verify_and_setup_llm(self) -> bool | None:
+		"""Verify and setup LLM (delegates to LLMService)."""
+		return self._llm_service.verify_and_setup_llm()
 
 
 _PythonAgent = Agent
